Loggers were added . Handled cases where the requested user detail wasn't present in DB in case of delete() and showUser()

+++ bootjpa/src/main/java/com/bootjpa/controller/UserController.java
@@ -2,6 +2,8 @@ package com.bootjpa.controller;
 
 import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -15,21 +17,27 @@ import com.bootjpa.dao.UserRepo;
 import com.bootjpa.model.User;
 import com.fasterxml.jackson.databind.util.JSONPObject;
 
+
 @Controller
 
 public class UserController {
+	
+	private final Logger logger =LoggerFactory.getLogger(this.getClass());
 	
 	@Autowired
 	UserRepo repo;
 
 	@RequestMapping("/")
 	public String home() {
+		logger.info("Page loaded successfully");
 		return "home.jsp";
+		
 	}
 	
 	@RequestMapping("/addUser")
 	public String addUser(User user) {
 		repo.save(user);
+		logger.info( user + " was added in the Table User");
 		return "home.jsp";
 	}
 	
@@ -48,17 +56,33 @@ public class UserController {
 	@RequestMapping("/showUser")
 	@ResponseBody
 	public   JSONPObject showUser(@RequestParam int userid) {
-		User user=repo.findById(userid).orElse(new User());
-		System.out.println("show user called");
-		return new JSONPObject("userobj", user);
+		boolean userexists=repo.existsById(userid);
+		if(userexists) {
+			User user=repo.findById(userid).orElse(new User());
+			logger.info("showUser function returned " + user + "  as JSON Object");
+			return new JSONPObject("userobj", user);
+		}
+		else {
+			logger.error("No such user exists with userid " + userid );
+			return new JSONPObject("userobj", "No user exists with userid "+userid);
+		}
+		
 	}
 	
 	
 	
 	@RequestMapping("/deleteUser")
 	public String deleteUser(@RequestParam int userid) {
+		boolean userexists= repo.existsById(userid);
+		if(userexists) {
 		repo.deleteById(userid);
-		System.out.println("user deleted");
+		logger.info("User with userid " +userid + " was deleted from Table User");
+		}
+		
+		else{
+		logger.error("No such user exists with userid " + userid);
+		}
+		
 		return "home.jsp";
 		
 	}
